--- Sidebar/src/App.css
@@ -1,115 +1,0 @@
-.App {
-  height: 100vh;
-  width: 100%;
-  padding: 0;
-  margin: 0;
-  display: flex;
-}
-
-.body {
-  margin: 0;
-  padding: 0;
-
-}
-
-/* Sidebar icons */
-
-.Sidebar {
-  height: 100%;
-  width: 20%;
-  background-color: rgb(44, 56, 70);
-  display: flex;
-  flex-direction: column;
-
-}
-
-.sails {
-  color: white;
-  padding-top: 10px;
-  height: auto;
-  display: flex;
-  justify-content: center;
-  font-size: 35px;
-}
-
-.Sidebardata {
-  height: auto;
-  padding: 0;
-  width: 100%;
-  margin-top: 20px;
-}
-
-.Sidebardata .row {
-  width: 100%;
-  height: 50px;
-  list-style-type: none;
-  margin: 0;
-  display: flex;
-  flex-direction: row;
-  color: white;
-  justify-content: center;
-  align-items: center;
-
-}
-
-.Sidebardata .row:hover {
-  cursor: pointer;
-  background-color: rgb(81, 80, 80);
-}
-
-.Sidebardata #active {
-  background-color: rgb(255, 218, 218);
-}
-
-.row #icon {
-  flex: 30%;
-  display: grid;
-  place-items: center;
-}
-
-.row #title {
-  flex: 70%;
-}
-
-/* Setting icons */
-
-.Setting {
-  height: auto;
-  padding: 0;
-  width: 100%;
-  margin-top: auto;
-
-
-
-}
-
-.Setting .row-s {
-  width: 100%;
-  height: 50px;
-  list-style-type: none;
-  margin: 0;
-  display: flex;
-  flex-direction: row;
-  color: white;
-  justify-content: center;
-  align-items: center;
-}
-
-.Setting .row-s:hover {
-  cursor: pointer;
-  background-color: rgb(81, 80, 80);
-}
-
-.Setting #active {
-  background-color: rgb(235, 220, 220);
-}
-
-.row-s #icon-s {
-  flex: 30%;
-  display: grid;
-  place-items: center;
-}
-
-.row-s #title-s {
-  flex: 70%;
-}
