--- Sidebar/src/App.js
@@ -1,16 +1,0 @@
-
-import './App.css';
-import Sidebar from './Components/Sidebar';
-import Admindashboard from './Components/Admindashboard';
-
-
-function App() {
-  return (
-    <div className="App">
-      <Sidebar />
-      <Admindashboard />
-    </div>
-  );
-}
-
-export default App;
