--- Sidebar/src/Components/Admindashboard.jsx
@@ -1,10 +1,0 @@
-import React from 'react';
-import '../App.css';
-
-
-function Admindashboard() {
-    return (
-        <div></div>
-    )
-}
-export default Admindashboard;
