+++ src/components/sidebar/index.jsx
@@ -2,17 +2,17 @@ import React from 'react';
 // import '../App.css';
 import { Sidebarlist } from './Sidebarlist';
 import { Setting } from './Sidebarlist';
-import './Sidebar.css';
+import styles from './index.module.scss';
 
 function Sidebar() {
     return (
-        <div className='Sidebar'>
-            <div className='sails'>Sails</div>
-            <ul className='Sidebardata'>
+        <div className={styles.sideBar}>
+            <div className={styles.sails}>Sails</div>
+            <ul className={styles.Sidebardata}>
                 {Sidebarlist.map((val, key) => {
                     return (
                         <li key={key}
-                            className='row'
+                            className={styles.row}
                             id={window.location.pathname === val.link ? "active" : " "}
                             onClick={() => { window.location.pathname = val.link; }}>
                             <div id='icon'>{val.icon}</div>
@@ -23,11 +23,11 @@ function Sidebar() {
             </ul>
 
 
-            <ul className='Setting'>
+            <ul className={styles.Setting}>
                 {Setting.map((val, key) => {
                     return (
                         <li key={key}
-                            className='row-s'
+                            className={styles["row-s"]}
                             id={window.location.pathname === val.link ? "active" : " "}
                             onClick={() => { window.location.pathname = val.link; }}>
                             <div id='icon-s'>{val.icon}</div>
