+++ src/screens/adminDashboard/index.jsx
@@ -1,7 +1,7 @@
 import React from 'react';
 // import './Components/Sidebar.css'
-import Sidebar from './Components/Sidebar';
-import './Admindashboard.css'
+import Sidebar from '../../components/sidebar';
+import styles from './index.module.scss'
 
 
 function Admindashboard() {
