+++ src/screens/loginPage/index.module.scss
@@ -1,0 +1,94 @@
+.main {
+    background-image: url("logo.jpeg");
+    background-size: contain;
+    justify-content: center;
+    display: flex;
+    padding: 90px 0 90px 0;
+}
+
+.sub-main {
+    display: flex;
+    justify-content: center;
+    height: 550px;
+    width: 35%;
+    box-shadow: 0px 8px 32px 0px rgb(46, 45, 45);
+    background-color: rgb(255, 255, 255, 0.15);
+    border-radius: 50px;
+    backdrop-filter: blur(8.5px);
+}
+
+.image {
+    display: flex;
+    justify-content: center;
+    padding: 20px;
+}
+
+.image-container {
+    background-color: gray;
+    border-radius: 150px;
+    align-items: center;
+    justify-content: center;
+    display: flex;
+    height: 115px;
+    width: 115px;
+}
+
+.profile {
+    height: 100px;
+    width: 100px;
+    display: flex;
+    border-radius: 130px;
+}
+
+.loginclass {
+    text-align: center;
+
+}
+
+.userimage {
+    height: 28px;
+    width: 28px;
+    position: absolute;
+    padding: 14px 0 0 25px;
+}
+
+
+
+.passwordimage {
+    height: 25px;
+    width: 25px;
+    position: absolute;
+    padding: 14px 0 0 25px;
+}
+
+input {
+    width: 300px;
+    height: 50px;
+    border-radius: 60px;
+    box-shadow: inset 0px 0px 25px 0px #8888;
+    padding-left: 80px;
+}
+
+.second-input {
+    padding-top: 20px;
+}
+
+.button {
+    width: 380px;
+    height: 50px;
+    border-radius: 60px;
+    background-color: rgb(90, 174, 237);
+    color: rgb(0, 0, 0);
+    font-size: 25px;
+    font-weight: bold;
+
+}
+
+.buttonclass {
+    padding-top: 30px;
+}
+
+.forgotpassword {
+    padding-top: 20px;
+    text-align: center;
+}
